requirements.txt
aiogram
django
djangorestframework
requests